generalize time since locked and generalize it with UI updates
