+++ client/src/app/components/environments/lock-time/lock-time.component.html
@@ -1,1 +1,13 @@
-<p-tag [value]="timeSinceLocked()" [rounded]="false" />
+<style>
+  .fixed-size-tag {
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    width: 100px; /* Set your desired width */
+    height: 30px; /* Set your desired height */
+    overflow: hidden; /* Ensure overflow content doesn't affect layout */
+    text-align: center; /* Center text within the tag */
+    white-space: nowrap; /* Prevent text wrapping */
+  }
+</style>
+<p-tag [value]="dateService.timeSinceLocked(this.lockedAt!, this.timeNow())" [rounded]="false" class="fixed-size-tag" />
