+++ client/src/app/pages/main-layout/main-layout.component.html
@@ -22,7 +22,7 @@
       @if (lockQuery.data()?.lockedAt) {
         <p class="text-sm text-gray-500 flex items-center gap-1 mb-5">
           <i-tabler name="lock"></i-tabler>
-          <span class="w-20">{{ timeSinceLocked(lockQuery.data()?.lockedAt) }}</span>
+          <span class="w-20">{{ dateService.timeSinceLocked(lockQuery.data()?.lockedAt, this.timeNow()) }}</span>
         </p>
       }
       <app-profile-nav-section />
