+++ client/src/app/pages/main-layout/main-layout.component.ts
@@ -14,6 +14,7 @@ import { DividerModule } from 'primeng/divider';
 import { ToastModule } from 'primeng/toast';
 import { TooltipModule } from 'primeng/tooltip';
 import { HeliosIconComponent } from '../../components/helios-icon/helios-icon.component';
+import { DateService } from '@app/core/services/date.service';
 
 @Component({
   selector: 'app-main-layout',
@@ -38,6 +39,7 @@ import { HeliosIconComponent } from '../../components/helios-icon/helios-icon.co
 export class MainLayoutComponent implements OnInit, OnDestroy {
   private keycloakService = inject(KeycloakService);
   private permissionService = inject(PermissionService);
+  dateService = inject(DateService);
 
   username = computed(() => (this.keycloakService.decodedToken()?.preferred_username || '') as string);
 
@@ -105,42 +107,6 @@ export class MainLayoutComponent implements OnInit, OnDestroy {
     // Clear the interval
     if (this.intervalId) {
       clearInterval(this.intervalId);
-    }
-  }
-
-  timeSinceLocked(lockedAt: string | undefined): string {
-    if (!lockedAt) return '';
-
-    const lockedDate = new Date(lockedAt);
-    const now = this.timeNow(); // your signal for "current time"
-    const diffMs = now.getTime() - lockedDate.getTime();
-
-    // If lockedAt is in the future, return empty or handle differently
-    if (diffMs < 0) {
-      return '';
-    }
-
-    const totalSeconds = Math.floor(diffMs / 1000);
-    const days = Math.floor(totalSeconds / 86400); // 1 day = 86400s
-    const hours = Math.floor((totalSeconds % 86400) / 3600);
-    const minutes = Math.floor((totalSeconds % 3600) / 60);
-    const seconds = totalSeconds % 60;
-
-    // Zero-pad hours, minutes, seconds
-    const padTwo = (num: number) => num.toString().padStart(2, '0');
-
-    if (days > 0) {
-      // Format: dd:hh:mm (e.g., "1:05:12" => 1 day, 5 hours, 12 minutes)
-      const dd = days.toString(); // or padTwo(days) if you prefer "01" for 1 day
-      const hh = padTwo(hours);
-      const mm = padTwo(minutes);
-      return `${dd}:${hh}:${mm}`;
-    } else {
-      // Format: hh:mm:ss (e.g., "05:12:36")
-      const hh = padTwo(hours);
-      const mm = padTwo(minutes);
-      const ss = padTwo(seconds);
-      return `${hh}:${mm}:${ss}`;
     }
   }
 }
